+++ tests.py
@@ -1,30 +1,34 @@
 import unittest
-from urlfix.urlfix import *
+from urlfix.urlfix import URLFix
+from urlfix.dirurlfix import DirURLFix
 import os
 import glob
 
 dir_path = os.path.dirname(os.path.abspath(__file__))
+dir_path = os.path.join("testfiles")
+replacement_file = os.path.join(dir_path,"replacement.txt")
 # Use the above to make paths to files, avoid changing directory just for tests.
+# ToDo: Create order in test files, add these to testfiles?
+# Todo: Avoid manually creating file paths.
 use_file = os.path.join(dir_path, "testurls.md")
 use_file_txt = os.path.join(dir_path, "testurls.txt")
-use_object = URLFix(input_file=use_file, output_file="replacement.txt")
-use_object_txt = URLFix(input_file=use_file_txt, output_file="replacement.txt")
+use_object = URLFix(input_file=use_file, output_file=replacement_file)
+use_object_txt = URLFix(input_file=use_file_txt, output_file=replacement_file)
 use_object_non_existent = URLFix(input_file=use_file, output_file="not_valid.txt")
-not_supported_files = URLFix(input_file=os.path.join(dir_path, "testurls.rst"), output_file="replacement.txt")
+not_supported_files = URLFix(input_file=os.path.join(dir_path, "testurls.rst"), output_file=replacement_file)
 use_object_inplace = URLFix(input_file=use_file)
 
-
-use_dir_object = DirURLFix(dir_path)
+use_dir_object = DirURLFix(os.path.join(dir_path, "testdir"))
 use_dir_non_existent = DirURLFix('non_existent')
 use_dir_non_dir = DirURLFix(use_file)
-use_files_dir = DirURLFix(os.path.join(dir_path, "testfiles"))
-use_files_dir_inplace = DirURLFix(os.path.join(dir_path, "testfiles"))
+use_files_dir = DirURLFix(os.path.join(dir_path, "testdir"))
+use_files_dir_inplace = DirURLFix(os.path.join(dir_path, "testdir"))
 
 
 class Testurlfix(unittest.TestCase):
     def test_instance_creation(self):
-        self.assertTrue(isinstance(use_object, URLFix))
-        self.assertTrue(isinstance(use_object_txt, URLFix))
+        [self.assertTrue(isinstance(urlfix_object, URLFix)) for urlfix_object in [use_object, use_object_txt]]
+
         with self.assertRaises(ValueError) as err:
             URLFix(input_file=use_file).replace_urls()
         self.assertEqual(str(err.exception), "Please provide an output file to write to.")
@@ -35,7 +39,8 @@ class Testurlfix(unittest.TestCase):
             use_object_non_existent.replace_urls()
         self.assertEqual(str(err.exception), "input_file and output_file should be valid files.")
         number_moved = use_object.replace_urls(verbose=0)
-        self.assertEqual(number_moved, 1)
+        # 3 out of 9 links should have moved.
+        self.assertEqual(number_moved, 3)
         number_moved_txt = use_object_txt.replace_urls(verbose=1)
         self.assertEqual(number_moved_txt, 2)
 
@@ -65,15 +70,17 @@ class TestDirURLFix(unittest.TestCase):
         self.assertEqual(str(err.exception), "File format rst is not yet supported.")
 
         # Check that if a known URL is provided, it is skipped
-        # Checking twice won't work since output files will exist already
-        number_moved_list = use_files_dir.replace_urls(correct_urls=["https://doi.org/10.5281/zenodo.3891106"],
-                                                       verbose=True)
+
+        number_moved_list = use_files_dir.replace_urls(
+            correct_urls=["https://zenodo.org/badge/DOI/10.5281/zenodo.3891106.svg"],
+            verbose=True)
         # Since we have three files, assert that the length returned is 3
         self.assertEqual(len(number_moved_list), 3)
 
-        self.assertEqual(number_moved_list[0], 1)
+        self.assertEqual(number_moved_list[0], 3)
         # 1 since we provided correct URLs. TODO: Figure out how to run both tests
-        self.assertEqual(number_moved_list[1], 1)
+        # 3 since we match double links if []()[]()
+        self.assertEqual(number_moved_list[1], 3)
         self.assertEqual(number_moved_list[2], 2)
         # Check skipping --> check that files are created in the above steps
         use_files_dir.replace_urls()
@@ -81,14 +88,17 @@ class TestDirURLFix(unittest.TestCase):
         self.assertTrue(os.path.isfile(os.path.join(dir_path, "testfiles", "testcorrect_output.md")))
         self.assertTrue(os.path.isfile(os.path.join(dir_path, "testfiles", "testurls_output.md")))
         self.assertTrue(os.path.isfile(os.path.join(dir_path, "testfiles", "txturls_output.txt")))
-        testfiles_path = os.path.join(dir_path, "testfiles")
+        # TODO: Automate file detection for unit tests.
+        testfiles_path = os.path.join(dir_path, "testdir")
         created_output_files = glob.glob(testfiles_path + "/*_output.*")
 
         for output_file in created_output_files:
-            try:
-                os.remove(output_file)
-            except:
-                pass
+            # Avoid try-except-else, not trivial to test
+            # Use this for now
+            # if this file exists and test passes, delete it
+            self.assertTrue(os.path.isfile(output_file))
+            print(f"Removing no longer needed file: {output_file}")
+            os.remove(output_file)
 
     def test_replace_urls_inplace(self):
         number_moved_list=use_files_dir_inplace.replace_urls(verbose=1, inplace=True)